ELEVATOR BASIC SYSTEM
Steps guide for ELEVATOR SYSTEM DESIGN :
ALGORITHIM BASED ON:
(FIFO) :First come First Out (Data Structure).
Requests is on a first-come-first-serve basis. The requests made by the passengers are added in a queue, as shown by the For simplicity, let’s assume that there are 11 floors in a building and a single elevator to serve requests. Suppose, a passenger on floor 2 wants to go down. This request is added to the queue. Next, a passenger on floor 4 wants to go up. This request is also added to the queue. More requests are added to the queue as they come in.